instruction set refactor
