+++ examples/_config.yml
@@ -2,4 +2,3 @@
 picture:
   source_image_dir: assets/images/_fullsize
   generated_image_dir: generated/
-  markup: picture
