+++ lib/jekyll-picture-tag/instruction_set.rb
@@ -1,6 +1,7 @@
+require 'yaml'
 # High level set of instructions for a given tag instance.
 class InstructionSet
-  attr_reader :source_images, :html_attributes
+  attr_reader :config, :preset
 
   def initialize(raw_tag_params, context)
     @context = context
@@ -14,23 +15,16 @@ class InstructionSet
     parse_tag_params raw_tag_params
   end
 
-  def preset_name
-    @preset_name || 'default'
+  def build_global_settings
+    default_config.merge @context.registers[:site]
   end
 
-  def media_presets
-    # Media queries
-    site.data['picture']['media-presets']
-  end
-
-  def preset
-    # Which preset we're using for this tag
-    site.data['picture']['markup-presets'][preset_name]
-  end
-
-  def output_format
-    # Which output format, such as <img> or <picture>
-    preset['markup'] || 'auto'
+  def build_preset
+    if site.data['picture']
+      default_preset.merge site.data['picture'][preset_name]
+    else
+      default_preset
+    end
   end
 
   # Returns the set of widths to use for a given media query.
@@ -85,78 +79,13 @@ class InstructionSet
     @context.registers[:site]
   end
 
-
   private
 
-  def parse_tag_params(raw_params)
-    # Raw argument example example:
-    # [preset] img.jpg [source_key: alt.jpg] [--(element || alt) attr=\"value\"]
-
-    # First, swap out liquid variables and split it on spaces into an array:
-    params = liquid_lookup(raw_params).split
-
-    # The preset is the first parameter, unless it's a filename.
-    # This regex is really easy to fool. TODO: improve it.
-    @preset_name = params.shift unless params.first =~ /[^\s.]+.\w+/
-
-    # source_image keys are media queries, values are source images. The first
-    # param specified will be our base image, so it has no associated media
-    # query.
-    @source_images = { nil => params.shift }
-
-    # Check if the next param is a source key, and if so assign it to the
-    # local variable source_key.
-    while params.first =~ /(?<media_query>\w+):/
-      params.shift # throw away the param, we already have the key
-      @source_images[media_query] = params.shift
-    end
-
-    # Anything left will be html attributes
-    build_html_attributes params.join(' ')
+  def default_config
+    Yaml.load File.read(__dir__ + '/defaults/global.yml')
   end
 
-  def liquid_lookup(params)
-    Liquid::Template.parse(params).render(@context)
-
-    # This gsub allows people to include template code for javascript libraries
-    # such as handlebar.js. It adds complication and I'm not sure it has much
-    # value now, so I'm commenting it out. If someone has a use case for
-    # it we can add it back in.
-    # .gsub(/\\\{\\\{|\\\{\\%/, '\{\{' => '{{', '\{\%' => '{%')
-  end
-
-  def build_html_attributes(params)
-    preset_attributes = preset['attributes'] || {}
-    @html_attributes = preset_attributes.transform_keys(&:to_sym)
-    # Example input:
-    # --picture class="awesome" --alt stumpy --img data-attribute="value"
-    params.split(' --').map(&:strip).each do |param|
-      # ['picture class="awesome"', 'alt stumpy', 'img data-attribute="value"']
-
-      # Splits on spaces, the first word will be our key.
-      a = param.split
-
-      # Supplied tag arguments will overwrite (not append) configured values
-      @html_attributes[a.shift.to_sym] = a.join(' ')
-    end
-  end
-
-  def site_url
-    # site url, example.com
-    site.config['url'] || ''
-  end
-
-  def site_path
-    # example.com/baseurl/
-    #            ^^^^^^^^^
-    site.config['baseurl'] || ''
-  end
-
-  def defaults
-    {
-      source_dir: '',
-      destination_dir: 'generated',
-      markup: 'auto'
-    }
+  def default_preset
+    Yaml.load File.read(__dir__ + '/defaults/preset.yml')
   end
 end
